+++ CHANGELOG.md
@@ -1,9 +1,12 @@
 CHANGELOG
 =========
 
+0.4.0
+-----
+
 * Added missing `timestamp` value to the typical statement.
 
-* when `null` is passed as the id argument to one of methods of the
+* when `null` is passed as the id argument to one of the methods of the
   `StatementFixtures` class, a unique UUID will be generated
 
 0.3.0